// src/app/interfaces/card-graph-stats.interface.ts
export interface CardGraphStats {
    headerClass: string;
    id: string;
    title: string;
    category: string;
    iconFooter: string;
    textFooter: string;
}

// <div class="card card-chart">
//     <div class="card-header card-header-success">
//         <div class="ct-chart" id="dailySalesChart"></div>
//     </div>
//     <div class="card-body">
//         <h4 class="card-title">Daily Sales</h4>
//         <p class="card-category">
//             <span class="text-success"><i class="fa fa-long-arrow-up"></i> 55% </span> increase in today sales.</p>
//     </div>
//     <div class="card-footer">
//         <div class="stats">
//             <i class="material-icons">access_time</i> updated 4 minutes ago
//         </div>
//     </div>
// </div>


// <div class="card card-chart">
//     <div class="card-header card-header-warning">
//         <div class="ct-chart" id="websiteViewsChart"></div>
//     </div>
//     <div class="card-body">
//         <h4 class="card-title">Email Subscriptions</h4>
//         <p class="card-category">Last Campaign Performance</p>
//     </div>
//     <div class="card-footer">
//         <div class="stats">
//             <i class="material-icons">access_time</i> campaign sent 2 days ago
//         </div>
//     </div>
// </div>